task_service/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
asyncpg==0.29.0
pydantic==2.6.4
sqlalchemy==2.0.23
pydantic_settings==2.2.1
passlib==1.7.4
PyJWT==2.9.0
